% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.vstar.R
\name{coef.vstar}
\alias{coef.vstar}
\title{Extract estimated parameters from VSTAR model}
\usage{
\method{coef}{vstar}(model, ...)
}
\arguments{
\item{model}{an object of S3-class \code{vstar}.}
}
\value{
List containing:
\itemize{
\item matrix \eqn{B}.
\item vector of \eqn{\gamma} values.
\item vector of threshold values.
}
}
\description{
Extract estimated parameters from VSTAR model
}
\keyword{internal}
