% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffn.R
\name{diffn}
\alias{diffn}
\title{Modfied \code{diff} function}
\usage{
diffn(x, lag = 1, differences = 1, na = NA, ...)
}
\arguments{
\item{x}{a vector or matrix with the values to be defferenced.
If a vector is provided it will be converted to a column matrix.}

\item{lag}{an integer indicating which lag to use.}

\item{differences}{an integer indicating the order of the difference.}

\item{na}{a value to fill missing points.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
a vector (!) with the computed differences.
}
\description{
Inspired by the function \code{diffn} from GAUSS.
Adds the value of \code{na} to the beginning of the result
to keep the initial size of the array.
}
\keyword{internal}
