% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.B.mat.R
\name{get.B.mat}
\alias{get.B.mat}
\title{Get coefficients estimates}
\usage{
get.B.mat(dataset, m, g, thr, g.function = "L")
}
\arguments{
\item{dataset}{an object of S3-classes \code{vstar.data} or \code{vstar}.}

\item{m}{a number of regimes in the model.}

\item{g}{a vector of smoothness parameter values.}

\item{thr}{a vector of threshold values.}

\item{g.function}{a transition function to be used
\itemize{
\item "L": Logistic transition function
\deqn{\frac{1}{1 + \exp (-\gamma (s_t - c))}}
\item "E": Exponential transition function
\deqn{1 - \exp(-\gamma (s_t - c)^2)}
\item fn: Custom function receiving transition variable, gammas and threshold
values in exactly this order.
}}
}
\value{
A list containing vectorized (!) matrix of estimated coefficients
\eqn{\operatorname{vec}\hat{B}} and the matrix of modified regressors
\eqn{M}.
}
\description{
An auxiliary function returning vectorized matrix of estimated coefficients
\eqn{\operatorname{vec}\hat{B}} and the matrix of modified regressors
\eqn{M}.

The code in this function is separated from the other code because all
lines of code are densely tied and enter the codebase as the
single chunk.

In this function implemented is the joint OLS estimation procedure from
(Teräsvirta and Yang, 2014).
}
\references{
T. Teräsvirta and Y. Yang,
“Specification, estimation and evaluation of
vector smooth transition autoregressive models with applications,”
Department of Economics and Business Economics, Aarhus University,
2014–08, Mar. 2014.
}
\keyword{internal}
