% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.bootstrap.R
\name{get.bootstrap}
\alias{get.bootstrap}
\title{sample values from a given set}
\usage{
get.bootstrap(x, n)
}
\arguments{
\item{x}{a data to sample from.}

\item{n}{a number of values to sample.}
}
\value{
A vector of sampled values.
}
\description{
sample values from a given set
}
\keyword{internal}
