% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.estimates.R
\name{get.estimates}
\alias{get.estimates}
\title{Get a set of estimated values}
\usage{
get.estimates(est, dataset, G.func)
}
\arguments{
\item{est}{a list containing the intermediate results of model estimation.
This list emerges as a result of inner loops in functions for estimating
VSTAR models (see \link{vstar.grid} and \link{vstar.nls}).}

\item{dataset}{an object of S3-classes \code{vstar.data} or \code{vstar}.}

\item{G.func}{a transition function to be used.
NB: it's a function, not its destription!}
}
\value{
A list containing
\itemize{
\item A matrix of estimated coefficients.
\item A matrix of estimated coefficients standard devations.
\item A matrix of estimated coefficients t-statistics.
\item A matrix of fitted values.
\item A matrix of residuals.
\item An estimate of a covariance matrix.
\item The value of \code{g.function}.
}
}
\description{
An auxiliary function returning a set of estimated values of a model.
Values are devectorized if needed.
}
\keyword{internal}
