% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.G.function.R
\name{get.G.function}
\alias{get.G.function}
\alias{get.G.derivative}
\title{Get transition function and its derivative}
\usage{
get.G.function(g.function = "L")

get.G.derivative(g.function = "L")
}
\arguments{
\item{g.function}{a transition function to be used
\itemize{
\item "L": Logistic transition function
\deqn{\frac{1}{1 + \exp (-\gamma (s_t - c))}}
\item "E": Exponential transition function
\deqn{1 - \exp(-\gamma (s_t - c)^2)}
\item fn: Custom function receiving transition variable, gammas and threshold
values in exactly this order.
}

If any custom function is provided to the \code{get.G.function} then
it's returned as-is.}
}
\description{
Auxiliary functions returning transition function and its derivative.
}
\keyword{internal}
