% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.grid.R
\name{get.grid}
\alias{get.grid}
\title{Get grid values}
\usage{
get.grid(g.limits, c.limits, points)
}
\arguments{
\item{g.limits}{a vector of lower and upper bounds of \eqn{\gamma}.}

\item{c.limits}{a vector of lower and upper bounds of \eqn{c}.}

\item{points}{a number of points in sequences.}
}
\value{
A list containing sequences of values of \eqn{\gamma} and \eqn{c}.
}
\description{
An auxiliary function to get two sequences of values of \eqn{\gamma}
and \eqn{c} in grid points.
}
\keyword{internal}
