% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.H.mat.R
\name{get.H.mat}
\alias{get.H.mat}
\title{Get the sequence of \eqn{H_{ij}} matrices}
\usage{
get.H.mat(B.mat)
}
\arguments{
\item{B.mat}{a matrix \eqn{\hat{B}}.}
}
\value{
A list containing the sequence of \eqn{H_{ij}} matrices.
}
\description{
An auxiliary function returning a sequence of \eqn{H_i} matrices
used in tests from (Teräsvirta and Yang, 2014).

The length of the sequence is exactly the number of elements in the
matrix of estimated coefficients \eqn{\hat{B}}.
Each matrix \eqn{H_{ij}} is a zero matrix with unity at the position
\eqn{(i, j)}. As we don't care about the ordering of these matrices
we order them in the order corresponding with R's inner representation
of matrices (i.e. up-to-down left-to-right).
}
\references{
T. Teräsvirta and Y. Yang,
“Linearity and misspecification tests for vector smooth transition
regression models,”
Department of Economics and Business Economics, Aarhus University,
2014–04, Feb. 2014.
}
\keyword{internal}
