% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.K.mat.R
\name{get.K.mat}
\alias{get.K.mat}
\title{Get matrix of the models first derivatives.}
\usage{
get.K.mat(model, J = 0)
}
\arguments{
\item{model}{an object of S3-class \code{vstar}.}

\item{J}{a number of first observations to be dropped.}
}
\value{
A matrix of first derivatives by \eqn{\{B, \Gamma, C\}}.
\eqn{\Omega} is excluded as all derivatives by it are zero.
}
\description{
An auxiliary function for estimating the matrix \eqn{K} of the first
derivatives of the model by its parameters \eqn{\{B, \Omega, \Gamma, C\}}.

This matrix is used in tests from (Teräsvirta and Yang, 2014).
}
\references{
T. Teräsvirta and Y. Yang,
“Linearity and misspecification tests for vector smooth transition
regression models,”
Department of Economics and Business Economics, Aarhus University,
2014–04, Feb. 2014.
}
\keyword{internal}
