% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagn.R
\name{lagn}
\alias{lagn}
\title{Modfied \code{lag} function}
\usage{
lagn(x, i, na = NA)
}
\arguments{
\item{x}{a vector or matrix with the values to be lagged
If a vector is provided it will be converted to a column matrix.}

\item{i}{an integer indicating which lag to use.}

\item{na}{a value to fill missing points.}
}
\value{
a vector (!) with the computed lags.
}
\description{
Inspired by the function \code{lagn} from GAUSS.
Adds the value of \code{na} to the beginning or the end of the result
to keep the initial size of the array.
}
\keyword{internal}
