% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinearity.test.R
\name{nonlinearity.test}
\alias{nonlinearity.test}
\title{Residual non-linearity test}
\usage{
nonlinearity.test(model, J = 1, stat.type = "all")
}
\arguments{
\item{J}{a number of lags of the transition variable to be included
in the test.}

\item{stat.type}{a string or a vector of strings setting the statistics
to be calculated and returned:
\itemize{
\item "all": calculate all statistics (default).
\item "LM": LM-type statistic only.
\item "r-LM": rescaled LM-statistic.
\item "Rao": Rao's F-statistic.
\item "Bartlett-Wilks": Bartlett's approximation of Wilks's \eqn{\Lambda}.
}}

\item{dataset}{an object of S3-classes \code{vstar.data} or \code{vstar}.}
}
\value{
A matrix containing the results in rows. Each row includes the value of
the corresponding statistic, 1\%, 5\%, and 10\% critical values, and
the p-value.
}
\description{
A test proposed by (Teräsvirta and Yang, 2014).
Based on the Tailor expansion of the transition function used in LM-type
test.

The null is that there is no residual non-linearity in the system.

As it's claimed in (Teräsvirta and Yang, 2014), LM-type statistic suffers
from size discrepancy that can lead to imcorrect results.
So three other statistics with improved size can be calculated as well:
\itemize{
\item Rescaled LM by (Leitinen, 1978; Meisner, 1979).
\item Rao's F (see, Rao, 1951).
\item Bartlett's approximation of Wilks's \eqn{\Lambda}-statistic
(see, Bartlett, 1954).
}
}
\references{
C. R. Rao,
“An asymptotic expansion of the distribution of Wilk’s criterion,”
Bulletin of the International Statistical Institute,
vol. 33, no. 2, Art. no. 2, 1951.

M. S. Bartlett,
“A Note on the Multiplying Factors for Various χ2 Approximations,”
Journal of the Royal Statistical Society. Series B (Methodological),
vol. 16, no. 2, pp. 296–298, 1954.

K. Laitinen,
“Why is demand homogeneity so often rejected?,”
Economics Letters,
vol. 1, no. 3, pp. 187–191, Jan. 1978.

J. F. Meisner,
“The sad fate of the asymptotic Slutsky symmetry test for large systems,”
Economics Letters,
vol. 2, no. 3, pp. 231–233, Jan. 1979.

T. Teräsvirta and Y. Yang,
“Linearity and misspecification tests for vector smooth transition
regression models,”
Department of Economics and Business Economics, Aarhus University,
2014–04, Feb. 2014.
}
