% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vstar.irf.R
\name{plot.vstar.irf}
\alias{plot.vstar.irf}
\title{Plotting Generalized Impulse Responses}
\usage{
\method{plot}{vstar.irf}(res, ...)
}
\arguments{
\item{res}{an object of S3-class \code{vstar.irf}.}
}
\description{
Plotting Generalized Impulse Responses
}
