% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.vstar.R
\name{predict.vstar}
\alias{predict.vstar}
\title{Model Predictions}
\usage{
\method{predict}{vstar}(model, new.data = NULL, new.trans = NULL, na.action = na.omit, ...)
}
\arguments{
\item{model}{an object of S3-classes \code{vstar}.}

\item{new.data}{if not NULL, use this new data as a source of variables for
prediction.}

\item{new.trans}{if not NULL, use this new transition variable for
prediction. Use this if the model was estimated with custom variable.}

\item{na.action}{an action that should be applied to NA values.}
}
\value{
A matrix of fitted.values.
}
\description{
The function returns the fitted values of the estimated model.
If \code{new.data} is not NULL then the fitted values for the new data are
calculated.
}
