% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.vstar.R
\name{summary.vstar}
\alias{summary.vstar}
\title{Printing the results of the model estimation}
\usage{
\method{summary}{vstar}(model, ...)
}
\arguments{
\item{model}{an object of S3-class \code{vstar}.}
}
\description{
Printing the results of the model estimation
}
\keyword{internal}
