% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unity.R
\name{unity}
\alias{unity}
\title{Get a unity vector of a specific legth}
\usage{
unity(n)
}
\arguments{
\item{n}{a number of elements.}
}
\value{
A column of \code{n} unities.
}
\description{
Get a unity vector of a specific legth
}
\keyword{internal}
