% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vstar-package.R
\docType{package}
\name{vstar-package}
\alias{vstar}
\alias{vstar-package}
\title{vstar: Functions and procedures for testing and estimating VSTAR models}
\description{
This package contains functions and procedures needed to test and estimate vector smooth transition autoregressive models (VSTAR). The codebase is based on two reports by Timo Teräsvirta and Yukai Yang (2014). Although the original papers describe models with different transition functions for every regime and every equation, in the package implemented is the simplest variant with the common transition function for each regime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vadim-zyamalov/vstar}
  \item \url{https://ideas.repec.org/p/aah/create/2014-04.html}
  \item \url{https://ideas.repec.org/p/aah/create/2014-08.html}
  \item Report bugs at \url{https://github.com/vadim-zyamalov/vstar/issues}
}

}
\author{
\strong{Maintainer}: Vadim Zyamalov \email{zyamalov@ranepa.ru}

Other contributors:
\itemize{
  \item Russian Presidential Academy of National Economic and Public Administration [copyright holder, funder]
}

}
\keyword{internal}
