% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vstar.girf.R
\name{vstar.girf}
\alias{vstar.girf}
\title{Generalized Impulse Response Functions}
\usage{
vstar.girf(
  model,
  shock = 1,
  n = 10,
  history = 1:model$dim$N,
  iter = 1000,
  R = 100,
  p.levels = c(0.1, 0.25, 0.5, 0.75, 0.9),
  cores = 1
)
}
\arguments{
\item{model}{an object of S3-classes \code{vstar}.}

\item{n}{a number of periods to calculate IRF.}

\item{history}{a range of observations, which forms the history before the
impulse range.}

\item{iter}{a number of steps to get the distribution of GIRF.}

\item{R}{a number of repetitions in each step.}

\item{p.levels}{a vector of needed percentiles in ascending order.
Five values should be provided.}

\item{cores}{a number of cores to be used.
If greater than 1 then a grid search will be done in parallel.
The number of cores has the upper bound of processor cores number minus 1.}

\item{shack}{a number of an equation, in which a shock appears.}
}
\value{
An object of S3 class \code{vstar.irf} containing
\itemize{
\item matrices of means of responses to positive and negative shocks.
\item matrices of corresponding standard deviations.
\item the number of the shock.
\item the length of the response range.
}
}
\description{
A function for constructing generalized impulse response functions using
Koop, et al. (1996) approach.
}
\details{
Exogenous variables are bootstrapped. \eqn{n} values of every exogenous
variable is sampled from \eqn{history} values. Determined components
are restored to continue the values from \eqn{history}.

New values of a transition variable depend on its nature.
\itemize{
\item If it's a lag of exogenous variable (maybe zero lag) then new values are
taken from old and new values of the corresponding exogenous variable.
\item If it's a lag of endogenous variable then new values are taken from
the existing ones, and after they are over the calculated values of the
corresponding variable are used.
\item If it's a custom variable then it's bootstrapped.
}

The error term is bootstrapped using the \eqn{history} values or the model
residuals.
}
\references{
G. Koop, M. H. Pesaran, and S. M. Potter,
“Impulse response analysis in nonlinear multivariate models,”
Journal of Econometrics, vol. 74, no. 1, pp. 119–147, 1996.
}
