% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vstar.nls.R
\name{vstar.nls}
\alias{vstar.nls}
\title{Improve VSTAR estimates using NLS}
\usage{
vstar.nls(model, tol = 1e-06, iter = 100, verbose = FALSE)
}
\arguments{
\item{model}{an object of S3-classes \code{vstar}.}

\item{tol}{a level of tolerance. If the vector of \eqn{\gamma} and \eqn{c}
changes less than this value the process stops.}

\item{iter}{the maximum number of iterations.}

\item{verbose}{should intermediate steps be shown?}
}
\value{
An object of S3-class \code{vstar}. See \link{vstar.grid} for details.
}
\description{
This function estimates the parameters of VSTAR model by NLS using
the provided \code{vstar} object as the source of data and starting values.
}
\details{
As stated by (Teräsvirta and Yang, 2014), estimating VSTAR directly by NLS
is quite hard due to the big number of parameters.
They proposed to estimate the model by grid search.
The results of this grid search can be used as the starting values for NLS.

The procedure goes as follows:
\itemize{
\item first, \eqn{\gamma} and \eqn{c} are reestimated using \link[stats]{optim}
with "Nelder-Mead" algorithm.
\item then, with these new values \eqn{B} is reestimated.
\item Repeat until convergence is reached.
}
}
\references{
T. Teräsvirta and Y. Yang,
“Specification, estimation and evaluation of
vector smooth transition autoregressive models with applications,”
Department of Economics and Business Economics, Aarhus University,
2014–08, Mar. 2014.
}
