% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vstar.prepare.R
\name{vstar.prepare}
\alias{vstar.prepare}
\title{Data preparation procedure}
\usage{
vstar.prepare(
  endo,
  exog = NULL,
  trans,
  const = FALSE,
  trend = FALSE,
  season = NULL,
  p = 1,
  coint.beta = NULL,
  coint.const = FALSE,
  coint.trend = FALSE,
  na.action = na.omit,
  dataset
)
}
\arguments{
\item{endo}{a vector of endogenous variables names.}

\item{exog}{a vector of exogenous variables names.}

\item{trans}{a transition variable. It can be either
\itemize{
\item a name of transition variable.
\item a list of 2 values, name of a variable and its lag, respectively.
\item a custom variable as a vector or a column matrix.
}}

\item{const}{should an intercept be included.}

\item{trend}{should a trend be included.}

\item{season}{should season dummies be included.
This parameter should be eigher NULL of interger number of seasons.}

\item{p}{a number of lags of autoregressive processes.}

\item{coint.beta}{an optional matrix of cointegrating vectors.
These should be pre-estimated.
This follows the idea from (Rothman, et al., 1999).
Cointegrating vectors should be located by rows.
Coefficients of constant term and trend (if any) should be the in the last
columns!}

\item{coint.const}{should an intercept be included into cointegration
relations.}

\item{coint.trend}{should a trend be included into cointegration relations.}

\item{na.action}{an action that should be applied to NA values.}

\item{dataset}{a data frame or a matrix with the data.}
}
\value{
An object of S3-class \code{vstar.data} containing three sublists.

The first one contains all input parameters except \code{p}, \code{na.action}, and
\code{dataset}.

The second one contains dimensions of the model: the final number of
observations, number of lags, and number of endogenous variables.

The last one conatins matrices of LHS \eqn{Y}, RHS \eqn{X},
trnsition variable \eqn{S}, and time \eqn{t}. The latter is not used at the
moment.
}
\description{
This function prepares data to be used in the estimation and testing
procedures.
}
\details{
All variables names should be either data frames or matrix column names.
}
\references{
P. Rothman, D. J. C. van Dijk, and P. H. B. F. Franses,
“A multivariate STAR analysis of the relationship between money and output,”
Erasmus University Rotterdam, Erasmus School of Economics (ESE),
Econometric Institute, EI 9945-/A, Nov. 1999.

T. Teräsvirta and Y. Yang,
“Specification, estimation and evaluation of
vector smooth transition autoregressive models with applications,”
Department of Economics and Business Economics, Aarhus University,
2014–08, Mar. 2014.
}
